Only send 'switch': 'on' if not already on.
This fixes the issues discovered here https://github.com/AlexxIT/SonoffLAN/issues/297#issuecomment-750379650.

Fill in the B05 scenes/effects

 as seen through the app

Remove harmfull kwargs checks

Remove the checks for a brightness/colour parameter in the kwargs. This prohibited switching to color or white from an effect. The checks are "duplicated" anyway in the get-or-default in the body of the ifs.