Improve Local Mode Reliability
Local mode devices cannot handle more than 1 concurrent connection
to the HTTP server. This means that if an additional connection is
made when another one is pending or is in the process of being
cleaned up by the device, it will return connection reset error.

==Mention of Failed Attempt==

An approach where we acquire a lock and wait 100ms whenever we
use the send() method was attempted but did not work consistently
on all devices. My theory is that depending the on the processing
on the device, it takes a different amount of time for it to
reallocate resources to the next new connection.

Therefore, simply retrying has worked just fine in my testing.
It also has the added advantage of working despite another HA
instance, app, etc. taking advantage of local mode.

Signed-off-by: rany <ranygh@riseup.net>